+++ django_afip/exceptions.py
@@ -1,6 +1,3 @@
-from . import parsers
-
-
 class AfipException(Exception):
     """
     Wraps around errors returned by AFIP's WS.
@@ -9,7 +6,7 @@ class AfipException(Exception):
     def __init__(self, response):
         Exception.__init__(self, 'Error {}: {}'.format(
             response.Errors.Err[0].Code,
-            parsers.parse_string(response.Errors.Err[0].Msg),
+            response.Errors.Err[0].Msg,
         ))