+++ CHANGELOG.rst
@@ -4,6 +4,12 @@ Changelog
 This file contains a brief summary of new features and dependency changes or
 releases, in reverse chronological order.
 
+2.5.0
+-----
+
+* We now rely on ``zeep``, rather ``suds``, update your dependencies
+  accordingly.
+
 2.4.0
 -----